# BERT CPU performance optimization: use mkldnn for nn.Linear() when input is dense layout