--- a/torch/utils/mkldnn.py
+++ b/torch/utils/mkldnn.py
@@ -27,7 +27,10 @@ class MkldnnLinear(torch.jit.ScriptModule):
 
     @torch.jit.script_method
     def forward(self, x):
-        return torch._C._nn.mkldnn_linear(x, self.weight, self.bias)
+        x_mkldnn = x if x.is_mkldnn else x.to_mkldnn()
+        y_mkldnn = torch._C._nn.mkldnn_linear(x_mkldnn, self.weight, self.bias)
+        y = y_mkldnn if x.is_mkldnn else y_mkldnn.to_dense()
+        return y
 
 
 class MkldnnConv2d(torch.jit.ScriptModule):