@mingfeima 2019-06-17
This PR assumes that `model` is converted to `mkldnn` as the following:
```python
from torch.utils import mkldnn as mkldnn_utils
model = mkldnn_utils.to_mkldnn(model)
```

I forked [BERT](https://github.com/huggingface/pytorch-pretrained-BERT) to create the benchmark, use `mkldnn_linear` branch from this [link](https://github.com/mingfeima/pytorch-pretrained-BERT/tree/mkldnn_linear) to reproduce result from this PR:
1. prepare dataset according to [link](https://gist.github.com/W4ngatang/60c2bdb54d156a41194446737ce03e2e).
2. update `GLUE_DIR` to actual dataset path in `run_inference.sh`.
3. test original performance, `bash run_inference.sh`.
4. test mkldnn linear performance, `bash run_inference.sh --mkldnn`.