@mingfeima 2019-06-17
TODOs:
`output` from `ideep` is not inplaced which will trigger memory copy by `mkldnn_to_dense`. Will update `ideep` interface to remove this copy. But the overall improvement is only marginal.