@Jianhui-Li 2019-06-18
@mingfeima @bddppq  

Mingfei's data is measured on a specific configuration where MKL-DNN has better performance than MKL. Let me communicate with MKL-DNN team how we should proceed  – whether we should replace MKL-DNN with MKL for all shapes/configuration,  or leave it as user option. 

This PR looses the condition of input format so that it take the dense input format so make the MKL-DNN integarted to the CPU path. I think it is better for us to stick to the principal that MKL-DNN integration to the MKL-DNN path.