@mingfeima 2019-06-24
Thanks for the review. Sorry for the late response, i have some family emergency to handle last week.

> so, you're trying to target only an explicitly converted model, right? we could also just prepend .to_mkldnn() somewhere in code before the model - would it be sufficient?

@dzhulgakov Yes, this is only targeting explicitly converted model. And i think aligns with current design.    And the model conversion serves as a switch whether you want to use `mkldnn` or not. This allows you to select best config flexibly (explained in section below).

Only prepending `input.to_mkldnn()` is not not sufficient as model conversion is used for `weight` caching for inference. And `input.to_mkldnn()` will trigger additional memory copy, this is not necessary since the model of BERT has no convolution so the `nn.Linear` is working on `plain` layout not `mkldnn` layout. So just do inplace (no memory copy) view between `at::tensor` and `ideep::tensor` will be enough.