@mingfeima 2019-06-24
> @mingfeima Could you clarify where does the perf gain come from? Is it:

@bddppq  First of all, i always compile with MKL...

The performance gain comes from that for some configs MKLDNN sgemm is faster than MKL's.

BERT has 3 major sizes of gemm: input_channel/output_channel = (768, 768), (768, 3072), (3072).
For the last two mkldnn is faster and for the first one mkl is faster, you can find the details in this [gist](https://gist.github.com/mingfeima/fe8110e8bbaaf0579e5d721ae3357697#impact-of-leading-dimension-padding).

GEMM performance is a huge topic and the current status is for some sizes mkldnn is faster and for others mkl is faster.
From Intel perspective, in case mkldnn has worse performance for some specific size, mkldnn team will try to fix it.
But this PR makes it possible to easily select the faster implementation for each particular size of the GEMMs even without Intel's support. You just have to:
1. collect hotspot gemm sizes;
2. write a benchmark to determine which is faster for a particular gemm, mkldnn or mkl?
3. in case using mkldnn is beneficial, convert this module, `mod = nn.Linear(ic, oc), mod.to_mkldnn()`; if not, leave it alone and mkl will be used.