@mingfeima 2019-06-26
@dzhulgakov @BIT-silence @bddppq This PR has been updated according to your feedback, please reivew.