@mingfeima 2019-07-04
just moved the `mkldnn`/`cpu` tensor conversion logic from `native/mkldnn/Linear.cpp` to `torch/utils/mkldnn.py`.

the conversion is `outplaced` at the moment and will increase some performance downgrade due to additional memory copy:

single socket result (20 cores) on Xeon 6148: (unit: iterations/sec):
1. mkl (original):
```bash
408/408 [00:24<00:00, 16.69it/s]
```
2. mkldnn: in place conversion, no memcpy:
```bash
408/408 [00:16<00:00, 24.06it/s]
```
3. mkldnn: out place conversion, no memcpy:
```bash
408/408 [00:18<00:00, 21.95it/s]
```