@mingfeima 2019-07-04
UPDATES: multiple instance result:
Xeon 6148: using 20 instances and 1 thread per each instance.
Benchmark has been rewritten with `ThroughputBenchmark`, use [run_inference.sh](https://github.com/mingfeima/pytorch-pretrained-BERT/blob/master/run_inference.sh) to reproduce.

1. before (mkl sgemm)
```bash
>>> ./run_inference.sh --multi_instances
Average latency per example: 86.198ms
Total number of iterations: 100
Total number of iterations per second (across all threads): 11.60
Total time: 8.620s
```

2. after (mkldnn sgemm)
```bash
>>> ./run_inference.sh --multi_instances --mkldnn
Average latency per example: 51.861ms
Total number of iterations: 100
Total number of iterations per second (across all threads): 19.28
Total time: 5.186s
```